A Machine Learning approach to detect Malicious Payloads in Android APKs
Smart Devices are become mandatory in human’s day to day life. According to
statisica.com, in India around 91% mobile users are Android users in 2019. There are
three security mechanisms are maintained by the company and the customer. 1.Device
Protection, 2. Data Protection and 3. Application Management. In which, Application
Management is the dif icult process by the company and the naïve users of the Android. In
this research paper we are proposing a hybrid approach to detect the malicious payloads.
This will be highly useful for the Android users. Based on the proposed approach, first we
use the SVM (Support Vector Machine) to classify the payload types. Secondly, we use the
LSTM(Long-Short Term Memory) method to detect the malicious payloads. The
experimental results proved the hybrid approach solves the payload detection in a easy
way.
Keywords: Android, Application Management, Malicious, Machine Learning
Approach.